Perf: drastically increased performance of DeleteGuestCustomers() method by using barebones DbContext

+++ src/Libraries/SmartStore.Services/Media/ImageCache.cs
@@ -111,7 +111,10 @@ namespace SmartStore.Services.Media
                 {
                     foreach (var file in _cacheRootDir.GetFiles())
                     {
-                        file.Delete();
+						if (!file.Name.IsCaseInsensitiveEqual("placeholder"))
+						{
+							file.Delete();
+						}
                     }
                     foreach (var dir in _cacheRootDir.GetDirectories())
                     {
